added helper method to describe better what the function of page.request from Playwright is
