+++ e2e/tests/admin/posts/post-settings.test.ts
@@ -1,10 +1,10 @@
 import {PostEditorPage, PostsPage} from '../../../helpers/pages';
 import {PostFactory, createPostFactory} from '../../../data-factory';
-import {expect, test} from '../../../helpers/playwright';
+import {expect, getHttpClient, test} from '../../../helpers/playwright';
 
 test.describe('Ghost Admin - Post - Settings', () => {
     test('shows correct publisher date format', async ({page}) => {
-        const postFactory: PostFactory = createPostFactory(page.request);
+        const postFactory: PostFactory = createPostFactory(getHttpClient(page));
         await postFactory.create({title: 'Test Post'});
 
         const postsPage = new PostsPage(page);
