+++ e2e/tests/post-factory.test.ts
@@ -1,12 +1,12 @@
 import {createPostFactory} from '../data-factory';
-import {expect, test} from '../helpers/playwright';
+import {expect, getHttpClient, test} from '../helpers/playwright';
 import type {PostFactory} from '../data-factory';
 
 test.describe('Post Factory API Integration', () => {
     let postFactory: PostFactory;
 
     test.beforeEach(async ({page}) => {
-        postFactory = createPostFactory(page.request);
+        postFactory = createPostFactory(getHttpClient(page));
     });
 
     test('create a post and view it on the frontend', async ({page}) => {
